requirements.txt
Flask
Flask-RESTx
Flask-SQLAlchemy
Flask-Migrate
flask-jwt-extended